virtio: fix return type of get_driver_features()
The virtio spec defines the features field to be 64 bits. The lower
parts of the virtio drivers (virtio::virtio_driver) treated it as such,
but the interface for each specific driver to override the feature bits
(virtio::virtio_driver::get_driver_features()) limited that to 32 bits
only. This patch rectifies it, changing the function's signature and all
overrides in the various drivers.

Message-Id: <AM0PR03MB62923C5820454DC1B9887D9AA6EF0@AM0PR03MB6292.eurprd03.prod.outlook.com>

+++ drivers/virtio-blk.cc
@@ -333,7 +333,7 @@ int blk::make_request(struct bio* bio)
     }
 }
 
-u32 blk::get_driver_features()
+u64 blk::get_driver_features()
 {
     auto base = virtio_driver::get_driver_features();
     return (base | ( 1 << VIRTIO_BLK_F_SIZE_MAX)
